+++ packages/procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/procaptcha-wrapper
 
+## 2.6.80
+### Patch Changes
+
+- Updated dependencies [bc3813d]
+- Updated dependencies [4d05e3f]
+  - @prosopo/types@4.4.0
+
 ## 2.6.79
 ### Patch Changes